# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/sopt/soptkathon/android/bckk/data/api/model/PostDisLikeRequest.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PostDisLikeRequest(
     @SerialName("postIdList")
-    val postIdList: List<Int>,
+    val postIdList: List<Long>,
 )
